Refactor GrindrPlus.kt to fix God Object anti-pattern
- Created `com.grindrplus.ui.DialogManager` to handle all dialog-related logic (version mismatch, bridge error, etc.).
- Created `com.grindrplus.core.NetworkRepository` to handle network requests (fetch remote data, fetch user ID).
- Refactored `com.grindrplus.core.Utils` to include `showToast` and `runOnMainThread` implementations.
- Updated `GrindrPlus.kt` to delegate logic to the new objects, significantly reducing its size and complexity.
- Updated internal usage of `showToast` in `Utils.kt`.
